refactor(sink): impl SinkFormatter for DebeziumJsonFormat
